Altered state data types from integer to long. This is expected by probes.
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Incubation/SimuLizar/trunk/org.palladiosimulator.simulizar@21070 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ src/org/palladiosimulator/simulizar/metrics/ResourceStateListener.java
@@ -91,7 +91,7 @@ public class ResourceStateListener implements IStateListener {
      * @see de.uka.ipd.sdq.simucomframework.resources.IStateListener#stateChanged(int, int)
      */
     @Override
-    public void stateChanged(final int queueLength, final int instanceId) {
+    public void stateChanged(final long queueLength, final int instanceId) {
         if (this.simuComModel.getSimulationControl().isRunning()) {
             final double simulationTime = this.simuComModel.getSimulationControl().getCurrentSimulationTime();
             if (this.lastTimeIdle) {
